requirements.txt
Flask==2.3.3
Werkzeug==3.0.6 
cryptography==43.0.1 
Flask-Talisman==1.0.0
mysql-connector-python==9.1.0
